pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"


[project]
name="physiolabxr"
version="0.1.0.dev1"
authors=[
    { name="Ziheng 'Leo' Li", email="s-vector.lee@hotmail.com"}
]
description="PhysioLabXR"

classifiers=[
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
    "brainflow",
    "numpy",
    "PyQt6",
    "pyserial",
    "pylsl",
    "scikit-learn",
    "scipy~=1.9.1",
    "pyqtgraph",
    "pyxdf",
    "pyscreeze",
    "opencv-python",
    "pyzmq",
    "setuptools",
    "psutil",
    "pytest-qt",
    "numba",
    "PyOpenGL",
    "PyOpenGL_accelerate",
    "pyaudio",
    "soundfile"
]

[project.urls]
"Homepage" = "https://github.com/physiolabxr/physiolabxr"
"Bug Tracker" = "https://github.com/physiolabxr/physiolabxr/issues"
"Documentation" = "https://physiolabxrdocs.readthedocs.io/en/latest/"

[project.scripts]
physiolabxr = "physiolabxr:physiolabxr"

[tool.setuptools]
include-package-data = true

#[tool.setuptools.packages.find]
#where = ["physiolabxr"]  # list of folders that contain the packages (["."] by default)
#include = ["physiolabxr*"]  # package names should match these glob patterns (["*"] by default)
#exclude = ["my_package.tests*"]  # exclude packages matching these glob patterns (empty by default)
#namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

#[tool.setuptools.package-data]
#package_dir = [
#    "_ui/*",
#    "_presets/*",
#    "_media/*",
#]